refactor: Implement Occasions feature and refactor product models
This commit introduces the Occasions feature, allowing users to view products based on a selected occasion. It also includes a significant refactoring of the product data structures and related test cases.

Key changes:
- Add `OccasionScreen` with a `TabBar` to switch between different occasions and display corresponding products.
- Implement `OccasionsCubit` to manage state for fetching products by occasion.
- Introduce `loading_indicator` for a better user experience during data fetching.
- Add test cases for `OccasionsCubit` to ensure correct state transitions for success and failure scenarios.
- Refactor product-related DTOs and entities (`ProductDto`, `ProductEntity`) for better code structure and reusability.
- Create a separate `PaginationDto` from the `ProductResponse`.
- Update data sources, repositories, use cases, and tests to align with the new product models.
- Add new translations for the Occasions screen.

+++ lib/core/app/presentation/widget/custom_card.dart
@@ -8,7 +8,7 @@ class CustomCard extends StatelessWidget {
   final String title;
   final double price;
   final double? oldPrice;
-  final int? discountPercentage;
+  final double? discountPercentage;
   final VoidCallback? onAddToCart;
   final VoidCallback? onTap;
